requirements.txt
elasticsearch
pandas
openai
beautifulsoup4
dspy
opentelemetry-api